ref: move handleAPIErrorResponse to util.go
